test: Move `normalizeProtoObject` to global utils
