+++ node/package.json
@@ -56,6 +56,7 @@
     "lodash": "^4.17.21",
     "log": "^6.3.1",
     "log-node": "^8.0.3",
+    "long": "^5.2.1",
     "memoizee": "^0.4.15",
     "mocha": "^9.2.2",
     "ncjsm": "^4.3.2",
