+++ node/packages/aws-lambda-sdk/test/lib/process-function.js
@@ -8,7 +8,7 @@ const { TracePayload } = require('@serverless/sdk-schema/dist/trace');
 const wait = require('timers-ext/promise/sleep');
 const basename = require('./basename');
 const awsRequest = require('../../../../test/utils/aws-request');
-const normalizeProtoObject = require('../utils/normalize-proto-object');
+const normalizeProtoObject = require('../../../../test/utils/normalize-proto-object');
 const resolveDirZipBuffer = require('../../../../test/utils/resolve-dir-zip-buffer');
 
 const fixturesDirname = path.resolve(__dirname, '../fixtures/lambdas');
