+++ node/test/utils/normalize-proto-object.js
